Split storage.rs
